requirements.txt
datasets>=3.0.0
pandas>=2.0.0
pyarrow>=15.0.0
pytest>=8.0.0
groq>=0.11.0
python-dotenv>=1.0.1
streamlit>=1.28.0
